На домашнюю добавлен список карточек
