+++ src/homeworks/homework-5/ProductList.tsx
@@ -1,29 +1,12 @@
 import classNames from 'classnames';
 import React, { useEffect, useState } from 'react';
-import { typeProduct } from 'src/homeworks/ts1/3_write';
 import { ProductItem } from 'src/stories/components/product/ProductItem';
-import '../../stories/components/product/product-list.scss';
 import Loader from 'src/stories/components/share/Loader';
-import { useStore } from 'src/store/state';
-import { createProductHelper } from 'src/common-functions/productHelper';
 import { TypeProduct } from 'src/types/typeProduct';
 import { useFetchProduct } from 'src/common-functions/useFetchProduct';
+import '../../stories/components/scss/product-list.scss';
 
 export type typeProductList = { listClssNames?: string; itemClssNames?: string };
-
-export interface ApiResponseProduct {
-  id: number;
-  title: string;
-  description: string;
-  price: number;
-  discountPercentage: number;
-  rating: number;
-  stock: number;
-  brand: string;
-  category: string;
-  thumbnail: string;
-  images: string[];
-}
 
 export const ProductList: React.FC<typeProductList> = (props) => {
   const { listClssNames = 'product-list', itemClssNames = 'product-item' } = props;
