+++ src/stories/components/header/Header.tsx
@@ -1,5 +1,10 @@
 import React from 'react';
-import './header.scss';
+import '../scss/header.scss';
+import { useUserAuth } from 'src/components/hooks/isAuth';
+import { useStore } from 'src/store/state';
+import { useGetExternalUser } from 'src/mock-data/mock-profile';
+import { NavLink } from 'react-router-dom';
+import { Logo } from 'src/stories/components/logo/Logo';
 
 type HeaderProps = {
   children?: React.ReactNode | React.ReactNode[];
@@ -10,5 +15,21 @@ export const Header = (props: HeaderProps) => {
     <header className="header">
       <div className="header-content">{props.children}</div>
     </header>
+  );
+};
+
+export const AppHeader = () => {
+  useGetExternalUser();
+  const isUserAuth = useUserAuth();
+  const loggedUser = useStore((state) => state.loggedUser);
+  return (
+    <Header>
+      <NavLink to={'/'}>
+        <Logo />
+      </NavLink>
+      <NavLink to={'products'}>Товары</NavLink>
+      <NavLink to={'bucket'}>Корзина</NavLink>
+      <NavLink to={'auth'}>{isUserAuth ? loggedUser?.username : `Вход`}</NavLink>
+    </Header>
   );
 };
