+++ src/stories/components/logo/Logo.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './logo.scss';
+import '../scss/logo.scss';
 
 export const Logo = () => {
   return <div className="logo">{/*<img src="../../assets/logo.jpg" alt="logo" className="logo-image" />*/}</div>;
