+++ src/stories/components/scss/bucket-item.scss
