+++ src/stories/components/scss/divider.scss
