+++ src/stories/components/scss/logo.scss
