+++ src/stories/components/scss/product-card.scss
@@ -2,6 +2,7 @@
     display: grid;
     justify-items: center;
     max-width: 300px;
+    min-width: 200px;
     max-height: 600px;
     padding: 1em;
     margin: 50px auto;
