+++ src/stories/components/scss/product-list.scss
