+++ src/stories/components/share/Divider.tsx
@@ -1,4 +1,4 @@
-import './divider.scss';
+import '../scss/divider.scss';
 
 export const Divider = () => {
   return <div className="divider"></div>;
